fix(approvals): notify approvers from the signer request-to-sign path
